Course final changes

+++ SimpleShoppingList/Controllers/ItemsEFController.cs
@@ -36,7 +36,7 @@ namespace SimpleShoppingList.Controllers
         }
 
         // PUT: api/ItemsEF/5
-        [ResponseType(typeof(void))]
+        [ResponseType(typeof(Item))]
         public IHttpActionResult PutItem(int id, Item item)
         {
             if (!ModelState.IsValid)
@@ -67,11 +67,11 @@ namespace SimpleShoppingList.Controllers
                 }
             }
 
-            return StatusCode(HttpStatusCode.NoContent);
+            return Ok(item);
         }
 
         // POST: api/ItemsEF
-        [ResponseType(typeof(Item))]
+        [ResponseType(typeof(ShoppingList))]
         public IHttpActionResult PostItem(Item item)
         {
             if (!ModelState.IsValid)
@@ -79,14 +79,24 @@ namespace SimpleShoppingList.Controllers
                 return BadRequest(ModelState);
             }
 
+            ShoppingList shoppingList = db.ShoppingLists
+                .Where(s => s.Id == item.ShoppingListId)
+                .Include(s => s.Items)
+                .FirstOrDefault();
+
+            if (shoppingList == null)
+            {
+                return NotFound();
+            }
+
             db.Items.Add(item);
             db.SaveChanges();
 
-            return CreatedAtRoute("DefaultApi", new { id = item.Id }, item);
+            return Ok(shoppingList);
         }
 
         // DELETE: api/ItemsEF/5
-        [ResponseType(typeof(Item))]
+        [ResponseType(typeof(ShoppingList))]
         public IHttpActionResult DeleteItem(int id)
         {
             Item item = db.Items.Find(id);
@@ -95,10 +105,21 @@ namespace SimpleShoppingList.Controllers
                 return NotFound();
             }
 
+            ShoppingList shoppingList = db.ShoppingLists
+                .Where(s => s.Id == item.ShoppingListId)
+                .Include(s => s.Items)
+                .FirstOrDefault();
+
+            if (shoppingList == null)
+            {
+                return NotFound();
+            }
+
+
             db.Items.Remove(item);
             db.SaveChanges();
 
-            return Ok(item);
+            return Ok(shoppingList);
         }
 
         protected override void Dispose(bool disposing)
